+++ src/nationalarchives/components/gallery/gallery.mjs
@@ -1,78 +1,86 @@
-import uuidv4 from "../../lib/uuid.mjs";
-
 export class Gallery {
   constructor($module) {
     this.$module = $module;
-    this.$items = $module && $module.querySelector(".tna-gallery__items");
-    this.opened = false;
+    this.$items = $module && $module.querySelectorAll(".tna-gallery__item");
+    this.$navigation =
+      $module && $module.querySelector(".tna-gallery__navigation");
+    this.$navigationItems =
+      this.$navigation &&
+      $module.querySelectorAll(".tna-gallery__navigation-item");
 
-    if (!this.$module || !this.$items) {
+    if (
+      !this.$module ||
+      !this.$items ||
+      !this.$navigation ||
+      !this.$navigationItems
+    ) {
       return;
     }
 
-    const supportTouchHorizontal = false;
-
-    this.items = this.$module.querySelectorAll(".tna-gallery__item").length;
-
-    if (this.items > 1) {
-      const uniqueId = `tna-gallery-${uuidv4()}`;
-
-      this.$module.classList.add("tna-gallery--collapsed");
-      this.$items.setAttribute("tabindex", "-1");
-
-      this.$galleryToggle = document.createElement("button");
-      this.$galleryToggle.classList.add(
-        "tna-gallery__toggle",
-        "tna-button",
-        "tna-button--accent",
-      );
-      this.$galleryToggle.setAttribute("aria-controls", uniqueId);
-      this.$galleryToggle.setAttribute("aria-expanded", false);
-      this.$galleryToggle.innerText = `View ${this.items} images`;
-
-      const onFirstTouch = () => {
-        this.$module.removeEventListener("touchstart", onFirstTouch);
-        this.$module.classList.add("tna-gallery--touchable");
-      };
-
-      if (supportTouchHorizontal) {
-        this.$module.addEventListener("touchstart", onFirstTouch);
-      }
-
-      this.$galleryToggle.addEventListener("click", () => {
-        if (supportTouchHorizontal) {
-          this.$module.removeEventListener("touchstart", onFirstTouch);
-        }
-        this.handleToggleGallery();
-      });
-
-      this.$galleryToggleWrapper = document.createElement("div");
-      this.$galleryToggleWrapper.classList.add(
-        "tna-column",
-        "tna-column--full",
-        "tna-gallery__toggle-wrapper",
-      );
-
-      this.$galleryToggleWrapper.appendChild(this.$galleryToggle);
-      this.$items.parentElement.appendChild(this.$galleryToggleWrapper);
-
-      this.$items.setAttribute("id", uniqueId);
-    }
+    this.setup();
+    this.currentId = this.$items[0].id;
+    this.showItem(this.currentId);
   }
 
-  handleToggleGallery() {
-    this.opened = !this.opened;
-    if (this.opened) {
-      this.$module.classList.remove("tna-gallery--collapsed");
-      this.$galleryToggle.setAttribute("aria-expanded", true);
-      this.$galleryToggle.innerText = "Close images";
-      this.$items.setAttribute("tabindex", "0");
-      this.$items.focus();
-      this.$items.setAttribute("tabindex", "-1");
-    } else {
-      this.$module.classList.add("tna-gallery--collapsed");
-      this.$galleryToggle.setAttribute("aria-expanded", false);
-      this.$galleryToggle.innerText = `View ${this.items} images`;
+  setup() {
+    this.$items.forEach(($item) => {
+      $item.setAttribute("hidden", "until-found");
+    });
+    this.$navigation.removeAttribute("hidden");
+    this.$navigationItems.forEach(($item) => {
+      $item.addEventListener("click", () =>
+        this.showItem($item.getAttribute("aria-controls")),
+      );
+    });
+    this.$module.addEventListener("keydown", (e) => {
+      switch (e.key) {
+        case "ArrowLeft":
+          this.showPreviousItem();
+          break;
+        case "ArrowRight":
+          this.showNextItem();
+          break;
+      }
+    });
+  }
+
+  showItem(id) {
+    this.$items.forEach(($item) => {
+      if ($item.id !== id) {
+        $item.setAttribute("hidden", "until-found");
+      } else {
+        $item.removeAttribute("hidden");
+      }
+    });
+    this.$navigationItems.forEach(($item) => {
+      if ($item.getAttribute("aria-controls") !== id) {
+        $item.setAttribute("aria-expanded", "false");
+      } else {
+        $item.setAttribute("aria-expanded", "true");
+      }
+    });
+    this.currentId = id;
+  }
+
+  getCurrentItemIndex() {
+    return Array.from(this.$items).findIndex(
+      ($item) => $item.id === this.currentId,
+    );
+  }
+
+  showPreviousItem() {
+    let nextIndexToShow = this.getCurrentItemIndex() - 1;
+    if (nextIndexToShow < 0) {
+      nextIndexToShow = this.$items.length - 1;
     }
+    this.showItem(this.$items[nextIndexToShow].id);
+  }
+
+  showNextItem() {
+    let nextIndexToShow = this.getCurrentItemIndex() + 1;
+    if (nextIndexToShow >= this.$items.length) {
+      nextIndexToShow = 0;
+    }
+    this.showItem(this.$items[nextIndexToShow].id);
   }
 }